requirements.txt
Flask==3.0.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
tensorflow==2.16.1
torch==2.2.1
torchvision==0.17.1
transformers==4.38.2
Pillow==10.2.0
matplotlib==3.8.3
dvc-s3